Intermediate State Refactoring
